LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/samesite/form-get-blank-reload.https-expected.txt

CONSOLE MESSAGE: [object MessageEvent]
CONSOLE MESSAGE: [object MessageEvent]
CONSOLE MESSAGE: [object MessageEvent]

PASS Reloaded same-host top-level form GETs are strictly same-site
PASS Reloaded subdomain top-level form GETs are strictly same-site
PASS Reloaded cross-site top-level form GETs are strictly same-site
